+++ generator/search_page_script.js
@@ -29,10 +29,6 @@ function is_on_desktop(){
 	return window.width > 1023;
 }
 
-if(is_on_desktop()){
-	document.getElementById("search_box").focus();
-}
-
 
 function on_search_input(){
 	const search_box = document.getElementById("search_box");
@@ -152,3 +148,38 @@ function on_search_input(){
 
 	results_div.innerHTML = results_div_body_str;
 }
+
+
+
+document.addEventListener("DOMContentLoaded", function() {
+	if(is_on_desktop()){
+		document.getElementById("search_box").focus();
+	}
+
+	if(location.hash !== ""){
+		const search_box = document.getElementById("search_box");
+
+		let search_str = "";
+
+		for(let i = 1; i < location.hash.length; i+=1){
+			if(location.hash[i] == '%'){
+				const char_code = parseInt(location.hash[i + 1]) * 16 + parseInt(location.hash[i + 2]);
+				search_str += String.fromCharCode(char_code);
+
+				i += 2;
+
+			}else{
+				search_str += location.hash[i];
+			}
+		}
+
+		search_box.value = search_str;
+
+		on_search_input();
+
+	}else{
+		const results_div = document.getElementById("results");
+		results_div.innerHTML = `<p style="padding-bottom: 1em; color: #878481;">Quick tip: this page will automatically search whatever is in the URL hash (add "#query" to end of URL)</p>`;
+	}
+});
+